Test against the real @graphiql/react instead of a hand mock
The plugin replaced the whole \`@graphiql/react\` module with a hand-written stub via a vitest alias, which silently drifts from the real package. Match the sibling plugins (\`doc-explorer\`, \`history\`): import the real module and \`vi.mock\` only the hooks (\`useGraphiQL\`, \`useGraphiQLActions\`), so icons, \`PanelHeader\`, and types come from the real package and stay honest.

A small \`graphiql-react-mock\` helper keeps the \`__state\` control surface the tests already used. Drops the module alias, adds the \`monaco-editor\` resolution alias and \`vi.mock(monaco-editor|zustand)\` the siblings use, and a \`use no memo\` directive in the test setup.
