+++ packages/graphiql-plugin-query-builder/src/components/__tests__/query-builder-variable-promotion.test.tsx
@@ -7,9 +7,15 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { GraphQLObjectType, GraphQLSchema, GraphQLString } from 'graphql';
-import { beforeEach, describe, expect, it } from 'vitest';
-import { __state } from '../../__mocks__/@graphiql/react';
+import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { __state, installGraphiQLReactMock } from './graphiql-react-mock';
 import { QueryBuilder } from '../query-builder';
+
+vi.mock('@graphiql/react', async () => {
+  const actual =
+    await vi.importActual<typeof import('@graphiql/react')>('@graphiql/react');
+  return { ...actual, useGraphiQL: vi.fn(), useGraphiQLActions: vi.fn() };
+});
 
 const QueryType = new GraphQLObjectType({
   name: 'Query',
@@ -35,6 +41,7 @@ describe('QueryBuilder — promoting a mutation argument to a variable', () => {
   let writes: { query?: string; variables?: string }[];
 
   beforeEach(() => {
+    installGraphiQLReactMock();
     writes = [];
     __state.schema = TestSchema;
     __state.queryText = 'mutation B {\n  setString(value: "hi")\n}';
