+++ packages/graphiql-plugin-query-builder/src/test-setup.ts
@@ -1,6 +1,8 @@
+'use no memo';
+
 import '@testing-library/jest-dom/vitest';
 import { cleanup } from '@testing-library/react';
-import { afterEach } from 'vitest';
+import { afterEach, vi } from 'vitest';
 
 // jsdom doesn't implement scrollIntoView; the cursor reveal calls it.
 Element.prototype.scrollIntoView ||= () => {};
@@ -8,3 +10,6 @@ Element.prototype.scrollIntoView ||= () => {};
 afterEach(() => {
   cleanup();
 });
+
+vi.mock('monaco-editor');
+vi.mock('zustand');
